Adds documentation diagrams and fixes bugs
Adds diagrams to documentation for database, blocks, and use cases to improve understanding.

Removes console log statements from backend and frontend code to clean up the codebase.

Pushes backend and frontend images to docker hub to make it easier to deploy.

Adds a required attribute to the genero field in the register form to improve data validation.